nix: init
Been messing with Nix and home-manager for a little over a week now.
It's pretty nice. I haven't run into any major problems yet, but there
are a few things I'm not satisfied with in regards to my current config.

TODO #1: Find out why I can't just add `./pkgs` to my `imports` list so
I don't have the ugly `_module.args.pkgs` stuff hanging around.

TODO #2: Find a way to link an aggregate folder to another folder (e.g.
`~/.nix-profile/share/fonts` to `~/.local/share/fonts`).

I'm not too keen on switching to NixOS yet, but only because it's the
middle of the semester and I'd rather have my computer in working order
for the time being. As soon as I have a lot of free time, I'll probably
run a backup and switch. But before that happens, I'm going to beef up
my home-manager config with basically everything I use, so it will be as
smooth a transition as possible. This will give me time to play with
other things, such as maybe setting up ZFS (encrypted) as the host FS,
and re-figuring out how to get VFIO running with NixOS.

Eventually I'll need to reorganize the repository so I can have my
home-manager config and system config in the same repo, but that day is
not today and so, for now, it'll just be my `~/.config/nixpkgs`.