flake: init
Switching over to flakes + nixus (I helped choose the name, so I'm still
riding that high). So far, going well-ish.

A little sad that I can't use my `toString ./secrets + "/${file}"` trick
anymore (since it gets the path of `./secrets` after the repo has been
imported into the store) -- need to find a way to not hardcode the path
to that folder (or at least, don't hardcode the username or path to
~/.config).

Also need to purify nixpkgs-mozilla (or hack-purify it, like colemickens
did with firefox-nightly-bin from that overlay) -- I want
firefox-beta-bin and stable Rust from it.