vin/wayland: switch to wezterm
It has built-in support for multiplexing and panes / tabs, which means
I can actually copy from a single pane without copying things in nearby
panes as well. It's also fairly easy to configure for the most part. I'll
Nix-ify my config once I'm happy with it (or not... I never did Nix-ify
my foot config either and it has been the same for a while now).