scadrial: update for rebuild
i9 13900k + RTX 4090, because I want good gaming performance, and Intel
was cheaper than AMD (and the only way I could get a 4090 was with a
bundle, which included an Intel motherboard, sadly).

If I don't blacklist the i915 module, my console will freeze after
a certain point in the boot process. Everything presumably works,
because I can SSH in, but... yeah. Need to enable nvidia modesetting and
blacklist i915 to get a TTY.

Also some hacks for getting the 4090 to work with sway, courtesy of
https://www.reddit.com/r/swaywm/comments/up1jvo/nvidia_driver_515_open_source_and_sway_no_more/
(archived:
https://web.archive.org/web/20220815103400/https://www.reddit.com/r/swaywm/comments/up1jvo/nvidia_driver_515_open_source_and_sway_no_more/).
Unfortunately, nouveau doesn't support AD102 chips yet, so I _have_
to use the closed-source driver for the moment. Maybe I'll look into
helping with that work...