vin/jj: make denylist include all children prior to merges
This way, all changes that have a parent that is denied (either wip
(disabled for now) or private) are also denied, except for merges. Because
my log colors denylisted changes differently, this makes it easier to
see when I wouldn't be able to push something because it depends on a
denylisted change.